refactor: move all utils functions out for better performance
